Renderer: fix delays in render state propagation
Separate render_tiles and render_sprites flags into undelayed and
delayed-by-one versions.

The render_tiles flag used for the palette selection (pal_index_delayed)
is out of sync with the propagated index value.  Need to fix this.

Also, the delay there is incorrect.  The delay should be RENDER_DELAY-2,
not RENDER_DELAY-1, because there is something in the pipeline causing an
extra delay (see comment in code).